server/requirements.txt
Django==4.0
djangorestframework==3.13.1
pytest-django==4.5.2
celery==5.1.2
redis==3.5.3